feat: moving the reducer into a top-level directory and out of `lib/`
